Add files via upload
In this project, we simulate the LTE with EPC network using NS-3. The simulation 
is done through three different types of applications and the information that describe
each flow like the Jitter sum or the number of lost packets, etc. all of them are plotted 
to see the differences between them through the applications.